% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paletas.R
\name{cores_ecoa}
\alias{cores_ecoa}
\title{Cores base da Ecoa Consultoria}
\format{
Um vetor de caracteres com 4 codigos hexadecimais:
\describe{
  \item{#4a3549}{Roxo escuro}
  \item{#aa6aef}{Roxo claro}
  \item{#f2aa84}{Pessego}
  \item{#FF8C00}{Laranja}
}
}
\usage{
cores_ecoa
}
\description{
Vetor com as 4 cores institucionais da Ecoa Consultoria Economica.
}
\examples{
cores_ecoa
plot(1:4, col = cores_ecoa, pch = 19, cex = 3)

}
