% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paletas.R
\name{cores_ecoa_cinza}
\alias{cores_ecoa_cinza}
\title{Cores da Ecoa com tons de cinza}
\format{
Um vetor de caracteres com 6 codigos hexadecimais:
\describe{
  \item{#4a3549}{Roxo escuro}
  \item{#aa6aef}{Roxo claro}
  \item{#948b9c}{Cinza arroxeado}
  \item{#cfc9c6}{Cinza claro}
  \item{#f2aa84}{Pessego}
  \item{#FF8C00}{Laranja}
}
}
\usage{
cores_ecoa_cinza
}
\description{
Variacao da paleta Ecoa com tons de cinza na transicao entre o roxo e o
laranja. Indicada para graficos com mais categorias, em que a paleta
original geraria cores intermediarias pouco distinguiveis.
}
\examples{
cores_ecoa_cinza
plot(1:6, col = cores_ecoa_cinza, pch = 19, cex = 3)

}
