% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paletas.R
\name{mostrar_paleta}
\alias{mostrar_paleta}
\title{Visualizar paleta discreta}
\usage{
mostrar_paleta(n)
}
\arguments{
\item{n}{Numero de cores a serem exibidas.}
}
\value{
Um objeto ggplot.
}
\description{
Cria um grafico mostrando n cores da paleta Ecoa com seus codigos hex.
}
\examples{
# Mostrar 4 cores
mostrar_paleta(4)

# Mostrar 8 cores
mostrar_paleta(8)

}
