% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paletas.R
\name{mostrar_paletas}
\alias{mostrar_paletas}
\title{Visualizar todas as paletas Ecoa}
\usage{
mostrar_paletas()
}
\value{
Um objeto ggplot.
}
\description{
Mostra todas as paletas do pacote em um unico grafico, no estilo de
\code{RColorBrewer::display.brewer.all()}.
}
\examples{
mostrar_paletas()

}
