% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paletas.R
\name{mostrar_paleta}
\alias{mostrar_paleta}
\title{Visualizar paleta discreta}
\usage{
mostrar_paleta(n, paleta = "ecoa", cinza = FALSE)
}
\arguments{
\item{n}{Numero de cores a serem exibidas.}

\item{paleta}{Nome da paleta (padrao: \code{"ecoa"}).}

\item{cinza}{Se TRUE, atalho para \code{paleta = "cinza"} (padrao: FALSE).}
}
\value{
Um objeto ggplot.
}
\description{
Cria um grafico mostrando n cores da paleta Ecoa com seus codigos hex.
}
\examples{
# Mostrar 4 cores da paleta padrao
mostrar_paleta(4)

# Mostrar 8 cores da paleta completa
mostrar_paleta(8, paleta = "completa")

}
