% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paletas.R
\name{paleta_ecoa}
\alias{paleta_ecoa}
\title{Funcao de paleta continua Ecoa}
\usage{
paleta_ecoa(n)
}
\arguments{
\item{n}{Numero de cores a serem geradas na paleta.}
}
\value{
Um vetor de caracteres com codigos hexadecimais das cores.
}
\description{
Gera uma funcao que interpola as cores da Ecoa para criar gradientes.
}
\examples{
# Gerar 10 cores
paleta_ecoa(10)

# Usar em um grafico base
barplot(1:8, col = paleta_ecoa(8))

}
