% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizecoa-package.R
\docType{package}
\name{vizecoa-package}
\alias{vizecoa-package}
\alias{vizecoa}
\title{vizecoa: Paleta de Cores Ecoa para ggplot2}
\description{
O pacote vizecoa fornece uma paleta de cores personalizada da Ecoa
Consultoria Economica e funcoes para integracao com ggplot2.
}
\section{Funcoes principais}{

\describe{
  \item{\code{\link{cores_ecoa}}}{Vetor com as 4 cores base da Ecoa}
  \item{\code{\link{cores_ecoa_cinza}}}{Variacao com tons de cinza, para mais categorias}
  \item{\code{\link{paleta_ecoa}}}{Funcao para gerar paletas continuas}
  \item{\code{\link{scale_fill_ecoa}}}{Escala de preenchimento continua}
  \item{\code{\link{scale_color_ecoa}}}{Escala de cor continua}
  \item{\code{\link{scale_fill_ecoa_d}}}{Escala de preenchimento discreta}
  \item{\code{\link{scale_color_ecoa_d}}}{Escala de cor discreta}
  \item{\code{\link{mostrar_paleta}}}{Visualizar a paleta}
  \item{\code{\link{hex_ecoa}}}{Obter codigos hex}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/moreirasd/viz_ecoa}
  \item Report bugs at \url{https://github.com/moreirasd/viz_ecoa/issues}
}

}
\author{
\strong{Maintainer}: Miguel Moreira \email{miguel@ecoaconsultoria.com}

Authors:
\itemize{
  \item Miguel Moreira \email{miguel@ecoaconsultoria.com}
}

}
\keyword{internal}
